refactor: rearrange Agenda props, and abstract internal components
