+++ src/components/Draggable.tsx
@@ -1,7 +1,7 @@
 import React from "react";
 import { GestureDetector } from "react-native-gesture-handler";
 import Animated from "react-native-reanimated";
-import { useDraggable } from "./hooks/useDraggable";
+import { useDraggable } from "../hooks/useDraggable";
 
 export const Draggable = ({ children }: { children: React.ReactNode }) => {
   const { dragGesture, animatedStyles } = useDraggable();
