--- src/types/appointment.ts
@@ -1,5 +1,0 @@
-export type Appointment = {
-  startDate: string;
-  endDate: string;
-  id: string;
-};
