+++ src/types/events.d.ts
@@ -1,0 +1,9 @@
+export type TEvent = {
+  startDate: string;
+  endDate: string;
+  id: string;
+};
+
+export type TAgendaData = Array<{
+  [date: string]: TAgenda[];
+}>;
